+++ src/redturtle/volto/restapi/serializer/summary.py
@@ -3,7 +3,6 @@ from plone import api
 from plone.app.contenttypes.interfaces import ILink
 from plone.app.contenttypes.utils import replace_link_variables_by_paths
 from plone.restapi.deserializer import json_body
-from plone.restapi.imaging import get_scale_infos
 from plone.restapi.interfaces import ISerializeToJsonSummary
 from plone.restapi.serializer.summary import (
     DefaultJSONSummarySerializer as BaseSerializer,
@@ -45,32 +44,6 @@ class DefaultJSONSummarySerializer(BaseSerializer):
         metadata_fields = query.get("metadata_fields", [])
         return "_all" in metadata_fields or self.force_all_metadata
 
-    def get_image_scales(self, data):
-        """
-        this is a backward compatibility for old volto templates that need
-        a full image scales object
-        """
-        metadata_fields = self.metadata_fields()
-
-        if "image" not in metadata_fields and not self.show_all_metadata_fields:
-            return {}
-        if data.get("image", None):
-            # it's a fullobjects data, so we already have the infos
-            return None
-        if not data.get("image_field", ""):
-            return None
-        scales = {}
-
-        for name, actual_width, actual_height in get_scale_infos():
-            scales[name] = {
-                "width": actual_width,
-                "height": actual_height,
-                "download": "{url}/@@images/{image_field}/{name}".format(
-                    url=data["@id"], image_field=data["image_field"], name=name
-                ),
-            }
-        return scales
-
     def get_remote_url(self):
         """
         Resolve uid and return ref absolute url
@@ -105,10 +78,6 @@ class DefaultJSONSummarySerializer(BaseSerializer):
             if v in EMPTY_STRINGS and k not in ["ExpirationDate", "EffectiveDate"]:
                 # this is a Volto compatibility
                 data[k] = None
-        scales = self.get_image_scales(data)
-
-        if scales:
-            data["image"] = {"scales": scales}
         if self.context.portal_type == "Link":
             remote_url = self.get_remote_url()
             # set twice because old templates can use both