+++ CHANGES.rst
@@ -131,7 +131,8 @@ Changelog
 
 - Max search limit became configurable by env var 'REDTURTLE_VOLTO_MAX_LIMIT_SEARCH', 500 by default.
   [folix-01]
-
+- Remove custom image infos in summary serializer. Now use plone.restapi base feature.
+  [cekk]
 
 5.2.2 (2023-08-29)
 ------------------